# Using a Consumer without consuming shouldn't throw a `RunloopTimeout` exception