Fixes #856 

We need this PR https://github.com/zio/zio-sbt/pull/220 to merged and a new release of zio-sbt to be made to fix the CI.

The solution implemented is to boot the Runloop lazily: we only start it when the user start a consuming session

## Benchmarks
### Before:
```scala
[info] Benchmark                                        Mode  Cnt     Score    Error  Units
[info] ConsumerBenchmark.throughput                     avgt   30   205.017 ±  6.977  ms/op
[info] ConsumerBenchmark.throughputWithCommits          avgt   30   284.384 ±  8.733  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients        avgt   30   647.721 ±  2.320  ms/op
[info] ConsumersComparisonBenchmark.manualKafkaClients  avgt   30   641.767 ±  2.508  ms/op
[info] ConsumersComparisonBenchmark.manualZioKafka      avgt   30  2673.759 ± 89.569  ms/op
[info] ConsumersComparisonBenchmark.zioKafka            avgt   30   737.250 ± 22.264  ms/op
```

### After:
```scala
[info] Benchmark                                        Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput                     avgt   30   209.078 ±   9.219  ms/op
[info] ConsumerBenchmark.throughputWithCommits          avgt   30   284.131 ±   9.911  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients        avgt   30   645.890 ±   3.103  ms/op
[info] ConsumersComparisonBenchmark.manualKafkaClients  avgt   30   642.727 ±   2.079  ms/op
[info] ConsumersComparisonBenchmark.manualZioKafka      avgt   30  1686.490 ± 219.173  ms/op
[info] ConsumersComparisonBenchmark.zioKafka            avgt   30   722.776 ±  38.728  ms/op
```
### Conclusions

Interestingly and unexpectedly it's making the manual consumption with zio-kafka kind of way faster 🤔

Here are the details about these changes in `manualZioKafka`:
Before:
```scala
sbt:zio-kafka> zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 5 -foe true .*ConsumersComparisonBenchmark.manualZioKafka*
...
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.manualZioKafka":
[info]   2655.102 ±(99.9%) 99.652 ms/op [Average]
[info]   (min, avg, max) = (1634.785, 2655.102, 2810.666), stdev = 201.302
[info]   CI (99.9%): [2555.450, 2754.754] (assumes normal distribution)
[info] # Run complete. Total time: 00:08:12
```

After:
```scala
sbt:zio-kafka> zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 5 -foe true .*ConsumersComparisonBenchmark.manualZioKafka*
...
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.manualZioKafka":
[info]   1706.941 ±(99.9%) 214.879 ms/op [Average]
[info]   (min, avg, max) = (1084.221, 1706.941, 2791.118), stdev = 434.067
[info]   CI (99.9%): [1492.062, 1921.820] (assumes normal distribution)
[info] # Run complete. Total time: 00:07:00
```

## Benchmarks after #889
### Before
(Results copied from #889)
```scala
zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 3 -foe true .*comparison.*
...
[info] Result "zio.kafka.bench.comparison.KafkaClientBenchmarks.kafkaClients":
[info]   641.438 ±(99.9%) 3.349 ms/op [Average]
[info]   (min, avg, max) = (624.941, 641.438, 650.032), stdev = 5.012
[info]   CI (99.9%): [638.090, 644.787] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.KafkaClientBenchmarks.manualKafkaClients":
[info]   642.514 ±(99.9%) 3.090 ms/op [Average]
[info]   (min, avg, max) = (632.069, 642.514, 656.932), stdev = 4.625
[info]   CI (99.9%): [639.424, 645.605] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaBenchmarks.manualZioKafka":
[info]   2605.392 ±(99.9%) 229.199 ms/op [Average]
[info]   (min, avg, max) = (1584.064, 2605.392, 2818.016), stdev = 343.055
[info]   CI (99.9%): [2376.193, 2834.592] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaBenchmarks.zioKafka":
[info]   734.987 ±(99.9%) 22.830 ms/op [Average]
[info]   (min, avg, max) = (683.740, 734.987, 826.528), stdev = 34.170
[info]   CI (99.9%): [712.158, 757.817] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaNoOptimisticResumeBenchmarks.manualZioKafka":
[info]   819.937 ±(99.9%) 120.082 ms/op [Average]
[info]   (min, avg, max) = (683.373, 819.937, 1243.105), stdev = 179.733
[info]   CI (99.9%): [699.855, 940.019] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaNoOptimisticResumeBenchmarks.zioKafka":
[info]   1311.588 ±(99.9%) 212.646 ms/op [Average]
[info]   (min, avg, max) = (755.915, 1311.588, 2329.325), stdev = 318.279
[info]   CI (99.9%): [1098.942, 1524.234] (assumes normal distribution)
...
[info] Benchmark                                            Mode  Cnt     Score     Error  Units
[info] KafkaClientBenchmarks.kafkaClients                   avgt   30   641.438 ±   3.349  ms/op
[info] KafkaClientBenchmarks.manualKafkaClients             avgt   30   642.514 ±   3.090  ms/op
[info] ZioKafkaBenchmarks.manualZioKafka                    avgt   30  2605.392 ± 229.199  ms/op
[info] ZioKafkaBenchmarks.zioKafka                          avgt   30   734.987 ±  22.830  ms/op
[info] ZioKafkaNoOptimisticResumeBenchmarks.manualZioKafka  avgt   30   819.937 ± 120.082  ms/op
[info] ZioKafkaNoOptimisticResumeBenchmarks.zioKafka        avgt   30  1311.588 ± 212.646  ms/op
```

### After
```scala
zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 3 -foe true .*comparison.*
...
[info] Result "zio.kafka.bench.comparison.KafkaClientBenchmarks.kafkaClients":
[info]   645.296 ±(99.9%) 3.731 ms/op [Average]
[info]   (min, avg, max) = (629.622, 645.296, 657.384), stdev = 5.585
[info]   CI (99.9%): [641.564, 649.027] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.KafkaClientBenchmarks.manualKafkaClients":
[info]   644.850 ±(99.9%) 3.046 ms/op [Average]
[info]   (min, avg, max) = (634.433, 644.850, 652.421), stdev = 4.559
[info]   CI (99.9%): [641.804, 647.896] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaBenchmarks.manualZioKafka":
[info]   1725.509 ±(99.9%) 283.478 ms/op [Average]
[info]   (min, avg, max) = (1083.098, 1725.509, 2688.894), stdev = 424.296
[info]   CI (99.9%): [1442.031, 2008.986] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaBenchmarks.zioKafka":
[info]   706.280 ±(99.9%) 23.099 ms/op [Average]
[info]   (min, avg, max) = (651.442, 706.280, 765.971), stdev = 34.573
[info]   CI (99.9%): [683.181, 729.379] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaNoOptimisticResumeBenchmarks.manualZioKafka":
[info]   891.336 ±(99.9%) 163.192 ms/op [Average]
[info]   (min, avg, max) = (556.650, 891.336, 1611.211), stdev = 244.258
[info]   CI (99.9%): [728.144, 1054.528] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.comparison.ZioKafkaNoOptimisticResumeBenchmarks.zioKafka":
[info]   1351.880 ±(99.9%) 459.663 ms/op [Average]
[info]   (min, avg, max) = (807.197, 1351.880, 4276.599), stdev = 688.002
[info]   CI (99.9%): [892.217, 1811.544] (assumes normal distribution)
...
[info] Benchmark                                            Mode  Cnt     Score     Error  Units
[info] KafkaClientBenchmarks.kafkaClients                   avgt   30   645.296 ±   3.731  ms/op
[info] KafkaClientBenchmarks.manualKafkaClients             avgt   30   644.850 ±   3.046  ms/op
[info] ZioKafkaBenchmarks.manualZioKafka                    avgt   30  1725.509 ± 283.478  ms/op
[info] ZioKafkaBenchmarks.zioKafka                          avgt   30   706.280 ±  23.099  ms/op
[info] ZioKafkaNoOptimisticResumeBenchmarks.manualZioKafka  avgt   30   891.336 ± 163.192  ms/op
[info] ZioKafkaNoOptimisticResumeBenchmarks.zioKafka        avgt   30  1351.880 ± 459.663  ms/op
```

### Conclusions

We still see this important improvement in the  `ZioKafkaBenchmarks.manualZioKafka` benchmark. The other ones are stable.