@erikvanoosten 2023-05-31
@guizmaii This is quite a sizeable refactoring. Could you guide us through the main ideas? Without those ideas it is hard to see if the implementation is correct. Some questions that pop are:

- Now that the run-loop time out is gone, when a consumer gets stuck processing, will we still stop polling from the broker?
- What is the purpose of using a Hub instead of what we had?
- There is a RunloopState.NotStarted and a RunloopState.Stopped, shall we merge them into a single state?

But please focus on the main ideas first.