@guizmaii 2023-06-04
@svroonland Done. Here's the PR where I do backported the piece of code which I was only refactoring: https://github.com/zio/zio-kafka/pull/901

I also rebase this branch on top of #901 so that this PR diff now only contains the code which is fixing #856