@guizmaii 2023-06-24
@erikvanoosten FYI, I remove the retry mechanism in the `RunloopAccess::stopConsumption` method. See https://github.com/zio/zio-kafka/pull/857/commits/75589412ff39a43bd187877a3966efdbf9e61110