@guizmaii 2023-06-24
FYI, I renamed `RunloopState.Stopped` to `RunloopState.Finalized` to make it clearer that a Runloop can't be restarted